lib: add GPS deframer
