+++ loragw/src/gps/mod.rs
@@ -1,2 +1,6 @@
 mod device;
+mod framing;
+mod types;
 pub use self::device::*;
+pub use self::framing::*;
+pub use self::types::*;
